test(cli): raise component coverage above 92 percent
